requirements.txt
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
python-multipart==0.0.6
PyNaCl==1.5.0
PyJWT==2.8.0
